requirements.txt
GitPython==3.1.27
gtrending==0.3.6
